% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_peloton_df.R
\name{augment_peloton_df}
\alias{augment_peloton_df}
\title{Add basic augmented fields to Peloton data frame}
\usage{
augment_peloton_df(df, tzone = "America/Los_Angeles")
}
\arguments{
\item{df}{Peloton data frame to process}

\item{tzone}{Name of time zone to convert the original UTC time from. Defaults to "America/Los_Angeles"}
}
\description{
This function augments a data frame obtained from get_workouts_df with some time and duration information.
}
\examples{
augment_peloton_df()
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{timezone,}
\keyword{workouts,}
