% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_peloton.R
\name{auth_peloton}
\alias{auth_peloton}
\title{POST's credentials to Pelton's auth/login endpoint to enable further requests}
\usage{
auth_peloton(login = Sys.getenv("PELOTON_LOGIN"),
  password = Sys.getenv("PELOTON_PASSWORD"))
}
\arguments{
\item{login}{Peloton Logi}

\item{password}{Peloton password}
}
\description{
POST's credentials to Pelton's auth/login endpoint to enable further requests
}
