% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with the Peloton API}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{Your Peloton screen name.}

\item{password}{Your Peloton password.}
}
\description{
This function allows you to authenticate with the Peloton API to pull your details and workout stats.
In this function you are able to pass in your `username` and `password`.  If you pass in no value,
the system will prompt you for a user name and password.
}
\examples{
authenticate()
authenticate("example user name", "example password")
}
\keyword{authenticate}
\keyword{peloton,}
