% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{combine_workout_data}
\alias{combine_workout_data}
\title{A helper function to join the individual workout dataframes}
\usage{
combine_workout_data(workouts_data)
}
\arguments{
\item{workouts_data}{a list of dataframes}
}
\value{
a dataframe
}
\description{
A helper function to join the individual workout dataframes
}
\examples{

}
