% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_all_workouts2}
\alias{get_all_workouts2}
\title{Deprecated: Use \code{\link{get_all_workouts}} instead}
\usage{
get_all_workouts2(limit_per_page = 100, max_pages = Inf, joins = "")
}
\arguments{
\item{limit_per_page}{Number of workouts to fetch per API request (default 100)}

\item{max_pages}{Maximum number of pages to fetch. Use \code{Inf} (default) to fetch all workouts.}

\item{joins}{Additional joins (e.g., "ride,ride.instructor")}
}
\description{
This function is deprecated. Use \code{\link{get_all_workouts}} which now
includes all the features of \code{get_all_workouts2()} including automatic
pagination and user ID auto-fetch.
}
\examples{
\dontrun{
# Use get_all_workouts() instead
get_all_workouts()
}

}
