% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ride_info_df.R
\name{get_basic_ride_info_df}
\alias{get_basic_ride_info_df}
\title{Get DataFrame with a Subset of Ride Information}
\usage{
get_basic_ride_info_df(ride_id)
}
\arguments{
\item{ride_id}{The id of the ride that you are pulling information for.}
}
\description{
This function returns a data frame with a subset of information related to a particular ride.
}
\examples{
get_basic_ride_info_df("55214456a1984c5885a087021e3f67b7")
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{ride,}
