% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructors_df.R
\name{get_instructors_df}
\alias{get_instructors_df}
\title{Get Data Frame with Instructors}
\usage{
get_instructors_df()
}
\description{
This function returns a data frame with all instructor information.
}
\examples{
get_instructors_df()
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{instructors,}
