% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_live_ride_and_details_df.R
\name{get_live_ride_and_details_df}
\alias{get_live_ride_and_details_df}
\title{Get Live Rides and Details}
\usage{
get_live_ride_and_details_df()
}
\description{
This function returns a data frame with all current live rides as well as the joined ride and instructor data for each ride.
}
\examples{
get_live_ride_and_details_df()
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{details,}
\keyword{instructors,}
\keyword{live,}
\keyword{ride,}
