% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_live_rides_df.R
\name{get_live_rides_df}
\alias{get_live_rides_df}
\title{Get Current Peloton Live Rides}
\usage{
get_live_rides_df()
}
\description{
This function returns a data frame with all of the current live rides.
}
\examples{
get_live_rides_df()
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{live,}
\keyword{ride,}
