% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_mapping_df.R
\name{get_metadata_mapping_df}
\alias{get_metadata_mapping_df}
\title{Get Metadata Mapping For Peloton Categories}
\usage{
get_metadata_mapping_df(type)
}
\arguments{
\item{type}{The type of metadata to pull. Available data types include: 'instructors', 'class_types', 'equipment', 'fitness_disciplines', device_type_display_names', 'content_focus_labels',}
}
\description{
This function returns a data frame of metadata available on Peloton for a particular category type.
}
\examples{
get_metadata_mapping_df("fitness_disciplines")
get_metadata_mapping_df("equipment")
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{metadata,}
