% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_stats_list.R
\name{get_my_stats_list}
\alias{get_my_stats_list}
\title{Get Full User Stats In a List}
\usage{
get_my_stats_list()
}
\description{
This function will gather all stats for the current authenticated user and return the results in a list.
}
\examples{
get_my_stats_list()

}
\keyword{Peloton,}
\keyword{authenticated,}
\keyword{list}
\keyword{stats,}
