% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_workout_stats_df.R
\name{get_my_workout_stats_df}
\alias{get_my_workout_stats_df}
\title{Get Workout Stats Data Frame}
\usage{
get_my_workout_stats_df()
}
\description{
This function will gather the workout counts by category for the current authenticated user.
}
\examples{
get_my_workout_stats_df()
}
\keyword{DataFrame}
\keyword{authenticated,}
\keyword{peloton,}
\keyword{stats,}
\keyword{workout,}
