% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{get_peloton_urls}
\alias{get_peloton_urls}
\title{Function for storing api urls}
\usage{
get_peloton_urls()
}
\arguments{
\item{NULL}{}
}
\value{
a list containing the Peloton API urls
}
\description{
Allows for storing API urls in one place for calling in other functions
}
\examples{
get_peloton_urls()

}
