% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{get_peloton_user_data}
\alias{get_peloton_user_data}
\title{Get User data from the Peloton API}
\usage{
get_peloton_user_data(userid)
}
\arguments{
\item{userid}{a string, the user id found in your Peloton profile}
}
\value{
a list, includes 3 data sets, base user data, workout counts and streaks
}
\description{
Includes data on User, workout counts, and workout streaks
}
\examples{
get_peloton_user_data(user_id)

}
