% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{get_peloton_workout_detail}
\alias{get_peloton_workout_detail}
\title{Return a data frame of workout data}
\usage{
get_peloton_workout_detail(userid, peloton_cookie)
}
\arguments{
\item{userid}{a string, the user id found in your Peloton profile}

\item{peloton_cookie}{a string, a cookie id}
}
\value{
a dataframe
}
\description{
Return a data frame of workout data
}
\examples{
userid <-  '5749258a9eb4389fbe23821'
peloton_cookie <- get_peloton_cookie(username, password)
get_peloton_workout_detail(userid, peloton_cookie)

}
