% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_perfomance_graphs}
\alias{get_perfomance_graphs}
\title{Makes a request against the \code{api/workout/workout_id/performance_graph} endpoint}
\usage{
get_perfomance_graphs(workout_ids, every_n = 5)
}
\arguments{
\item{workout_ids}{WorkoutIDs}

\item{every_n}{How often measurements are reported. If set to 1, there will be 60 data points per minute of a workout.}
}
\description{
For each workout, returns time series of individual workouts capturing cadence, output, resistance, speed, heart-rate (if applicable), measured at second intervals defined by \code{every_n}. A vectorized function, so accepts multiple \code{workoutIDs} at once.
}
\examples{
\dontrun{
workouts <- get_all_workouts()
get_perfomance_graphs(workouts$id)
}

}
