% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_performance_graphs}
\alias{get_performance_graphs}
\title{Makes a request against the \code{api/workout/workout_id/performance_graph} endpoint}
\usage{
get_performance_graphs(
  workout_ids,
  every_n = 5,
  dictionary = list(list = c("seconds_since_pedaling_start", "segment_list")),
  parse_dates = TRUE,
  ...
)
}
\arguments{
\item{workout_ids}{WorkoutIDs}

\item{every_n}{How often measurements are reported. If set to 1, there will be 60 data points per minute of a workout}

\item{dictionary}{A named list mapping a data-type to a column name}

\item{parse_dates}{Whether to try and guess which columns are dates and convert}

\item{...}{Other arguments passed on to methods}
}
\description{
For each workout, returns time series of individual workouts capturing cadence, output, resistance, speed, heart-rate (if applicable), measured at second intervals defined by \code{every_n}. A vectorized function, so accepts multiple \code{workoutIDs} at once.
}
\examples{
\dontrun{
workouts <- get_all_workouts()
get_performance_graphs(workouts$id)
get_performance_graphs(workouts$id,
  dictionary =
    list("list" = c("seconds_since_pedaling_start", "segment_list"))
)
}

}
