% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ride_info_list.R
\name{get_ride_info_list}
\alias{get_ride_info_list}
\title{Get Full Set of Ride Information In a List}
\usage{
get_ride_info_list(ride_id)
}
\arguments{
\item{ride_id}{The id of the ride that you are pulling information for.}
}
\description{
This function returns a list with all information related to a particular ride.
}
\examples{
get_ride_info_list('55214456a1984c5885a087021e3f67b7')
}
\keyword{Peloton,}
\keyword{list}
\keyword{ride,}
