% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rider_information.R
\name{get_rider_information}
\alias{get_rider_information}
\title{Requests complete Rider, Instructor, and Workout Information}
\usage{
get_rider_information(
  userid = Sys.getenv("PELOTON_USERID"),
  num_workouts = 20,
  tz = "America/New_York"
)
}
\arguments{
\item{userid}{userID}

\item{num_workouts}{Number of workouts requested}

\item{tz}{Timezone to convert times into}
}
\value{
A tibble of all rider, instructor, and workout information
}
\description{
Requests complete Rider, Instructor, and Workout Information
}
