% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_workout_data}
\alias{get_workout_data}
\title{Makes a request against the \code{api/workout/workout_id} endpoint}
\usage{
get_workout_data(
  workout_id,
  date_parsing = TRUE,
  dictionary = list(numeric = c("v2_total_video_watch_time_seconds",
    "v2_total_video_buffering_seconds", "leaderboard_rank"), list =
    c("achievement_templates"))
)
}
\arguments{
\item{workout_id}{WorkoutID}

\item{date_parsing}{Whether to try and guess which columns are dates and convert}

\item{dictionary}{A named list. Maps a data-type to a column name. If \code{NULL} then no parsing is done}
}
\description{
Returns data about individual workouts. A vectorized function, so accepts multiple \code{workoutIDs} at once.
}
\examples{
\dontrun{
get_workout_data(
  workout_id = workout_id,
  dictionary = list(
    "numeric" = c(
      "v2_total_video_watch_time_seconds", "v2_total_video_buffering_seconds",
      "v2_total_video_watch_time_seconds", "leaderboard_rank"
    ),
    "list" = c("achievement_templates")
  )
)
}
}
