% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{get_workout_detail_json}
\alias{get_workout_detail_json}
\title{Returns the json for the workout detail, requires authentication}
\usage{
get_workout_detail_json(workout_url, peloton_cookie)
}
\arguments{
\item{workout_url}{a string, generated from the get_workout_url() function}

\item{peloton_cookie}{a string, cookie id}
}
\value{
JSON
}
\description{
Returns the json for the workout detail, requires authentication
}
\examples{

}
