% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{get_workout_stream_url}
\alias{get_workout_stream_url}
\title{Get API url for a workout's stream data}
\usage{
get_workout_stream_url(workout_id, time_offset_interval = 1)
}
\arguments{
\item{workout_id}{a string of and individial workout id}

\item{time_offset_interval}{}
}
\value{
the user workout stream API call
}
\description{
Get API url for a workout's stream data
}
\examples{
workout_id <- '44e4b4f4c14434444a4f554444a44bde4'
get_workout_stream_url(user_id)

}
