% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{get_workout_url}
\alias{get_workout_url}
\title{Get API url for a specific user id}
\usage{
get_workout_url(userid)
}
\arguments{
\item{userid}{a string, the user id found in your Peloton profile}
}
\value{
the user workout API call
}
\description{
Get API url for a specific user id
}
\examples{
user_id <- '44e4b4f4c14434444a4f554444a44bde4'
get_workout_url(user_id)

}
