% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workouts_and_instructors_df.R
\name{get_workouts_and_instructors_df}
\alias{get_workouts_and_instructors_df}
\title{Get Workout and Instructor Details in a DataFrame}
\usage{
get_workouts_and_instructors_df(user_id)
}
\arguments{
\item{user_id}{Your Peloton user id.}
}
\description{
This function will gather all individual workouts and all related instructor data for a user and
return it in a data frame. If a user id is not passed into the function, it will return the current users workouts.
}
\examples{
get_workouts_and_instructors_df()
}
\keyword{DataFrame}
\keyword{Peloton,}
\keyword{authenticated,}
\keyword{instructors,}
\keyword{workouts,}
