% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_workouts_data}
\alias{get_workouts_data}
\title{Makes a request against the \code{api/workout/workout_id} endpoint}
\usage{
get_workouts_data(workout_ids)
}
\arguments{
\item{workout_ids}{WorkoutIDs}
}
\description{
A vectorized function that returns data about individual workouts.
}
\examples{
\dontrun{
peloton_auth() ; get_all_workouts()
}

}
