% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{parse_base_user_data}
\alias{parse_base_user_data}
\title{Parses user json to return base user data}
\usage{
parse_base_user_data(json_user)
}
\arguments{
\item{json_user}{JSON, json return from the user API call}
}
\value{
a dataframe
}
\description{
Parses user json to return base user data
}
\examples{
user_id '5749258a9eb4389fbe23821'
urls <- get_peloton_urls()
json_user <- jsonlite::fromJSON(paste0(urls$base.url, "user/", userid))
parse_base_user_data(json_user)

}
