% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{parse_credentials_to_cookie}
\alias{parse_credentials_to_cookie}
\title{Helper function that parses the cookie id from JSON}
\usage{
parse_credentials_to_cookie(credentials)
}
\arguments{
\item{credentials}{JSON, the returned json from the API after authentication}
}
\value{
a string, cookie id
}
\description{
Helper function that parses the cookie id from JSON
}
\examples{
parse_credentials_to_cookie(credentials)

}
