% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_list_to_df.R
\name{parse_dates}
\alias{parse_dates}
\title{Helper function to infer which columns are dates (UNIX epoch) and parse them as such.}
\usage{
parse_dates(dataframe, tz = "America/Los_Angeles")
}
\arguments{
\item{dataframe}{A dataframe containing some columns that may be dates}

\item{tz}{Timezone to convert datestamp to}
}
\description{
Automatically converts timezone to America/Los_Angeles
}
\examples{
\dontrun{
parse_dates(data.frame(a = 1570914652, b = "adad", c = 123L))
}

}
