% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{parse_ride_stats}
\alias{parse_ride_stats}
\title{Returns ride stats data frame}
\usage{
parse_ride_stats(workouts_json)
}
\arguments{
\item{workouts_json}{a string, generated from the get_workout_url() function}
}
\value{
a dataframe
}
\description{
Returns ride stats data frame
}
\examples{

}
