% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{parse_user_workouts}
\alias{parse_user_workouts}
\title{Returns user workouts data frame}
\usage{
parse_user_workouts(workouts_json)
}
\arguments{
\item{workouts_json}{a string, generated from the get_workout_url() function}
}
\value{
a dataframe
}
\description{
Returns user workouts data frame
}
\examples{

}
