% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peloton_apis}
\alias{peloton_apis}
\title{Useful API endpoints}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{endpoint}{the endpoint to query against}
  \item{optional_args }{if you can augment your query with additional arguments}
  \item{description}{description about what the data contains}
  ...
}}
\usage{
peloton_apis
}
\description{
A dataset containing Peloton API endpoints and some useful metadata about them
}
\keyword{datasets}
