% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_auth.R
\name{peloton_auth}
\alias{peloton_auth}
\title{Deprecated Peloton authentication helper}
\usage{
peloton_auth(...)
}
\arguments{
\item{...}{Ignored. Authentication now relies on the \code{PELOTON_BEARER_TOKEN} environment variable.}
}
\description{
Deprecated Peloton authentication helper
}
\keyword{internal}
