% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{peloton_user_id}
\alias{peloton_user_id}
\title{Makes a call to \code{\link{get_my_info}}}
\usage{
peloton_user_id(...)
}
\arguments{
\item{...}{Other arguments passed on to methods}
}
\description{
Returns \code{userid}.
}
\examples{
\dontrun{
peloton_user_id()
}

}
