% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{update_types}
\alias{update_types}
\title{Update data types after parsing to dataframe}
\usage{
update_types(df, dictionary = NULL)
}
\arguments{
\item{df}{the output of parse_list_to_df}

\item{dictionary}{dictionary to interpret}
}
\description{
Sometimes certain columns are adadeada
}
\examples{
\dontrun{
df <- dplyr::tibble(
  z = c("1", "2", "3"),
  x = c(9, 8, 7), f = "a"
)
dictionary <- list(
  "numeric" = c("x", "z"),
  "character" = c("m", "f")
)
update_types(df, dictionary)
}

}
