% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_peloton.R
\name{auth_peloton}
\alias{auth_peloton}
\title{POST's credentials to Pelton's auth/login endpoint to enable further requests}
\usage{
auth_peloton()
}
\description{
POST's credentials to Pelton's auth/login endpoint to enable further requests
}
