% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_all_workouts}
\alias{get_all_workouts}
\title{Makes a request against the \code{api/user_id/workouts/} endpoint}
\usage{
get_all_workouts(userid = Sys.getenv("PELOTON_USERID"), num_workouts = 20)
}
\arguments{
\item{userid}{userID}

\item{num_workouts}{num_workouts}
}
\description{
Lists requested number of workouts for a user, along with some metadata.
}
\examples{
\dontrun{
peloton_auth()
get_all_workouts()
}

}
