% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_my_info}
\alias{get_my_info}
\title{Makes a request against the \code{api/me} endpoint}
\usage{
get_my_info()
}
\description{
Returns user metadata, including userid, email, account status, etc..
}
\examples{
\dontrun{
peloton_auth() ; get_my_info()
}

}
