% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_my_info}
\alias{get_my_info}
\title{Makes a request against the \code{api/me} endpoint}
\usage{
get_my_info(dictionary = NULL, parse_dates = TRUE, ...)
}
\arguments{
\item{dictionary}{A named list. Maps a data-type to a column name. If \code{NULL} then no parsing is done}

\item{parse_dates}{Whether to try and guess which columns are dates and convert}

\item{...}{Other arguments passed on to methods}
}
\description{
Returns user metadata, including userid, email, account status, etc.  \code{userid} is particularly useful since you need it for \code{\link{get_workouts_data}}.
}
\examples{
\dontrun{
peloton_auth()
get_my_info()
}

}
