% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_my_info}
\alias{get_my_info}
\title{Makes a request against the \code{api/me} endpoint}
\usage{
get_my_info(dictionary = NULL, ...)
}
\arguments{
\item{dictionary}{a dict}

\item{...}{arguments passed onto methods}
}
\description{
Returns user metadata, including userid, email, account status, etc.  \code{userid} is particularly useful since you need it for \code{\link{get_workouts_data}}.
}
\examples{
\dontrun{
peloton_auth()
get_my_info()
}

}
