% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_perfomance_graphs}
\alias{get_perfomance_graphs}
\title{Makes a request against the \code{api/workout/workout_id/performance_graph} endpoint}
\usage{
get_perfomance_graphs(workout_ids, every_n = 5)
}
\arguments{
\item{workout_ids}{WorkoutIDs}

\item{every_n}{Everywhat?}
}
\description{
Returns time series of individual workouts capturing cadence, output, resistance and speed measured at many intervals.
}
\examples{
\dontrun{
workouts <- get_all_workouts()
get_perfomance_graphs(workouts$id)
}

}
