% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_workouts_data}
\alias{get_workouts_data}
\title{Makes a request against the \code{api/workout/workout_id} endpoint}
\usage{
get_workouts_data(
  workout_ids,
  dictionary = list(numeric = c("v2_total_video_watch_time_seconds",
    "v2_total_video_buffering_seconds", "v2_total_video_watch_time_seconds",
    "leaderboard_rank"), list = c("achievement_templates")),
  parse_dates = TRUE,
  ...
)
}
\arguments{
\item{workout_ids}{WorkoutIDs}

\item{dictionary}{A named list mapping a data-type to a column name}

\item{parse_dates}{Whether to try and guess which columns are dates and convert}

\item{...}{Other arguments passed on to methods}
}
\description{
Returns data about individual workouts. A vectorized function, so accepts multiple \code{workoutIDs} at once.
}
\examples{
\dontrun{
get_workouts_data(
  workout_ids = workout_ids,
  dictionary = list(
    "numeric" = c(
      "v2_total_video_watch_time_seconds", "v2_total_video_buffering_seconds",
      "v2_total_video_watch_time_seconds", "leaderboard_rank"
    ),
    "list" = c("achievement_templates")
  )
)
}
}
