% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_dates}
\alias{parse_dates}
\title{Convert UNIX epoch timestamps to datetime}
\usage{
parse_dates(dataframe, tz = base::Sys.timezone())
}
\arguments{
\item{dataframe}{A dataframe containing some columns that may be dates}

\item{tz}{Timezone to convert datestamp to}
}
\description{
Helper function convert UNIX timestamps to datestamps. By default converts to \code{America/Los_Angeles} timezone.
}
\examples{
\dontrun{
parse_dates(data.frame(a = 1570914652, b = "adad", c = 123L))
}

}
