% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_list_to_df.R
\name{parse_list_to_df}
\alias{parse_list_to_df}
\title{Parsing the JSON content of a response and turning result into a tibble. Converts nested lists into list-columns.}
\usage{
parse_list_to_df(list)
}
\arguments{
\item{list}{The JSON content of a response (a named list in R)}
}
\description{
Parsing the JSON content of a response and turning result into a tibble. Converts nested lists into list-columns.
}
