% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_list_to_df}
\alias{parse_list_to_df}
\title{Parse the \code{JSON} content of a response and turn the result into a \code{tibble} .}
\usage{
parse_list_to_df(list, parse_dates, dictionary)
}
\arguments{
\item{list}{The JSON content of a response (aka a named list in R)}

\item{parse_dates}{Whether to turn epoch timestamps into datetimes}

\item{dictionary}{A dict}
}
\description{
A helper, agnostic function to parse the content of API responses made to Peloton's API. Leaves most datatypes alone, but converts nested lists into list-columns.
}
\examples{
\dontrun{
parse_list_to_df(peloton_api("api/me")$content)
}

}
