% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_list_to_df.R
\name{parse_list_to_df}
\alias{parse_list_to_df}
\title{Parse the \code{JSON} content of a response and turning result into a tibble.}
\usage{
parse_list_to_df(list, parse_dates = TRUE)
}
\arguments{
\item{list}{The JSON content of a response (a named list in R)}

\item{parse_dates}{Whether to turn timestamps into datetimes}
}
\description{
Helper agnostic function to parse the content for API responses made to Peloton's API. Leaves most datatypes alone, but converts nested lists into list-columns.
}
\examples{
\dontrun{
parse_list_to_df(peloton_api("api/me")$content)
}

}
