% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{peloton_api}
\alias{peloton_api}
\title{Makes a \code{GET} request against one of Peloton's API endpoints}
\usage{
peloton_api(path, ...)
}
\arguments{
\item{path}{API endpoint to query}

\item{...}{Additional parameters passed onto methods}
}
\description{
Users need not invoke this method directly and may instead use one of the wrappers around specific endpoints, such as
}
\examples{
\dontrun{
peloton_auth()
peloton_api("api/me")
}

}
