% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_api.R
\name{peloton_api}
\alias{peloton_api}
\title{Makes a GET request against one of Peloton's APIs}
\usage{
peloton_api(path, ...)
}
\arguments{
\item{path}{API endpoint to query}

\item{...}{Additional parameters passed onto methods}
}
\description{
Makes a GET request against one of Peloton's APIs
}
