% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peloton_apis}
\alias{peloton_apis}
\title{Useful API endpoints}
\format{A data frame with 4 rows and 4 variables:
\describe{
  \item{endpoint}{the endpoint to query against}
  \item{optional_args }{arguments you can augment your query with}
  \item{description}{description about what the response data contains}
  ...
}}
\usage{
peloton_apis
}
\description{
A dataset containing Peloton API endpoints and some useful metadata about them
}
\keyword{datasets}
