% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_auth.R
\name{peloton_auth}
\alias{peloton_auth}
\title{\code{POST} credentials to Peloton's auth/login endpoint}
\usage{
peloton_auth(login = Sys.getenv("PELOTON_LOGIN"),
  password = Sys.getenv("PELOTON_PASSWORD"))
}
\arguments{
\item{login}{Peloton login}

\item{password}{Peloton password}
}
\description{
Users needs to run this each session in order to authenticate any requests made against other Pelton endpoints. Set environmental variables \code{`PELOTON_LOGIN`} and \code{`PELOTON_PASSWORD`}, or provide them yourself.
}
\examples{
\dontrun{
peloton_auth() # need to run each session to authorize any other API requests
}
}
