% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloton_auth.R
\name{peloton_auth}
\alias{peloton_auth}
\title{POST's credentials to Pelton's auth/login endpoint to enable further requests}
\usage{
peloton_auth(login = Sys.getenv("PELOTON_LOGIN"),
  password = Sys.getenv("PELOTON_PASSWORD"))
}
\arguments{
\item{login}{Peloton login}

\item{password}{Peloton password}
}
\description{
User needs to set environmental variables `PELOTON_LOGIN` and `PELOTON_PASSWORD` to authenticate API requests
}
\examples{
\dontrun{
peloton_auth() # need to run each session to authorize any other API requests
}
}
